Crate shn

shn-rs

provides an api to work with *.shn files found in the game Fiesta Online Said files are statically typed data tables with a less then ideal format.

This library contains methods for easily reading and writing those files, you should parse them into a different format for actual use though, since the structures are not very ergonomic for actually using the data.

Structs

ShnColumn
Represents a column in the SHN table.

ShnFile
Represents a SHN file

ShnRow
Represents a single row of data within a file

ShnSchema
Represents the schema of an shn file, which is defined by a collection of ShnRows

Enums

ShnCell
Represents a single data cell within the SHN-File

ShnDataType
Represents a data type within a SHN File.

ShnError
Wrapper for errors within the shn-rs crate.

Functions

read_from
Reads a ShnFile from the provided input, using the given encoding for any strings

write_to
Writes the ShnFile to the provided output, using the given encoding for any strings.