Struct shn::ShnColumn

pub struct ShnColumn {
    pub name: String,
    pub data_type: ShnDataType,
    pub data_length: i32,
}

Represents a column in the SHN table.

Fields

name

The name identifing the column

data_type

The type of data being held by this column

data_length

The length of the data being held in this column. Only relevant for string types.

Methods

impl ShnColumn

fn new_string_fixed_len(name: &str, len: i32) -> Self

Constructs a new column with the StringFixedLen type

fn new_string_terminated(name: &str) -> Self

Constructs a new column with the StringZeroTerminated type

fn new_byte(name: &str) -> Self

Constructs a new column with the Byte type

fn new_signed_byte(name: &str) -> Self

Constructs a new column with the SignedByte type

fn new_unsigned_short(name: &str) -> Self

Constructs a new column with the UnsignedShort type

fn new_signed_short(name: &str) -> Self

Constructs a new column with the SignedShort type

fn new_unsigned_integer(name: &str) -> Self

Constructs a new column with the UnsignedInteger type

fn new_signed_integer(name: &str) -> Self

Constructs a new column with the SignedInteger type

fn new_single_floating_point(name: &str) -> Self

Constructs a new column with the SingleFloatingPoint type